requirements.txt
gradio==4.44.0
google-genai==0.3.0
python-dotenv==1.0.1